More cleanup and adding unit test case
